提升版本

+++ src/main/java/com/gregtechceu/gtceu/api/gui/widget/IntInputWidget.java
@@ -48,7 +48,7 @@ public class IntInputWidget extends NumberInputWidget<Integer> {
 
     @Override
     protected String toText(Integer value) {
-        return String.valueOf(value);
+        return String.valueOf(value.intValue());
     }
 
     @Override
